+++ src/math/core/tools/random/Xorshift.js
@@ -9,23 +9,6 @@
  */
 
 /**
- * Multiply two 32-bit integers and output a 32-bit integer.
- * @param {number} x1 
- * @param {number} x2 
- * @returns {number}
- * @private
- * @ignore
- */
-const multiplication32 = function(x1, x2) {
-	let y = ((x1 & 0xFFFF) * (x2 & 0xFFFF)) >>> 0;
-	let b = (x1 & 0xFFFF) * (x2 >>> 16);
-	y = (y + ((b & 0xFFFF) << 16)) >>> 0;
-	b = (x1 >>> 16) * (x2 & 0xFFFF);
-	y = (y + ((b & 0xFFFF) << 16));
-	return (y & 0xFFFFFFFF);
-};
-
-/**
  * Random number class.
  * @private
  * @ignore
@@ -84,9 +67,9 @@ export default class Xorshift {
 	setSeed(seed) {
 		// seedを使用して線形合同法で初期値を設定
 		let random_seed = seed;
-		random_seed = (multiplication32(random_seed, 214013) + 2531011) >>> 0;
+		random_seed = (Math.imul(random_seed, 214013) + 2531011) >>> 0;
 		this.z = random_seed;
-		random_seed = (multiplication32(random_seed, 214013) + 2531011) >>> 0;
+		random_seed = (Math.imul(random_seed, 214013) + 2531011) >>> 0;
 		this.w = random_seed;
 
 		/**